refactor: refactor dialogs
