+++ src/features/create-todo/ui/create-todo.module.scss
@@ -7,6 +7,6 @@
 .form {
   display: flex;
   flex-direction: column;
-  gap: 1rem;
+  gap: 1.5rem;
   padding-block: 8px;
 }
