+++ src/features/update-todo/ui/update-todo-dialog.tsx
@@ -1,38 +1,60 @@
-import styles from './create-todo.module.scss';
-import { FormEvent, memo } from 'react';
+import { FormEvent, memo, useEffect } from 'react';
+import { TodoDialog } from '@/shared/ui/todo-dialog';
+import styles from '@/features/create-todo/ui/create-todo.module.scss';
 import { useForm } from '@tanstack/react-form';
 import { TodoSchema } from '@/entities/todo/model';
 import { useTodoStore } from '@/shared/store/todos.store';
-import { TodoDialog } from '@/shared/ui/todo-dialog';
 import { useDialogStore } from '@/shared/store/dialog.store';
+import { format } from 'date-fns';
 
-export const CreateTodo = memo(function CreateTodo() {
-  const addTodo = useTodoStore((state) => state.addTodo);
+export const UpdateTodoDialog = memo(function UpdateTodo() {
+  const updateTodo = useTodoStore((state) => state.updateTodo);
+  const getTodo = useTodoStore((state) => state.getTodo);
   const closeDialog = useDialogStore((state) => state.closeDialog);
+  const selectedTodoId = useDialogStore((state) => state.selectedTodoId);
 
   const form = useForm({
     defaultValues: {
       title: '',
       dueDate: new Date().toDateString(),
     },
-    onSubmit: ({ value }) => {
-      addTodo({ ...value, dueDate: new Date(value.dueDate) });
-      closeDialog();
-    },
     validators: {
       onChange: TodoSchema,
     },
+    onSubmit({ value }) {
+      if (selectedTodoId) {
+        updateTodo({
+          ...value,
+          dueDate: new Date(value.dueDate),
+          id: selectedTodoId,
+        });
+        closeDialog();
+      }
+    },
   });
 
+  useEffect(() => {
+    if (selectedTodoId) {
+      const todo = getTodo(selectedTodoId);
+      if (todo) {
+        form.setFieldValue('title', todo.title);
+        form.setFieldValue(
+          'dueDate',
+          format(new Date(todo.dueDate), 'yyyy-MM-dd'),
+        );
+      }
+    }
+  }, [form, form.state, getTodo, selectedTodoId]);
+
   return (
-    <TodoDialog headline={'Создать задачу'} dialogName={'create'} fullscreen>
+    <TodoDialog dialogName={'update'} headline={'Редактировать задачу'}>
       <form
         onSubmit={(e) => {
           e.preventDefault();
           form.handleSubmit();
         }}
         className={styles['form']}
-        id={'createTodoForm'}
+        id={'updateTodoForm'}
       >
         <form.Field name={'title'}>
           {(field) => (
@@ -68,16 +90,16 @@ export const CreateTodo = memo(function CreateTodo() {
       </form>
 
       <mdui-button
-        variant="outlined"
+        variant="text"
         slot="action"
         onClick={closeDialog}
         type="reset"
-        form={'createTodoForm'}
+        form={'updateTodoForm'}
       >
         Отмена
       </mdui-button>
-      <mdui-button slot="action" type="submit" form={'createTodoForm'}>
-        Создать
+      <mdui-button slot="action" type="submit" form={'updateTodoForm'}>
+        Сохранить
       </mdui-button>
     </TodoDialog>
   );
