+++ src/shared/store/dialog.store.ts
@@ -7,13 +7,13 @@ interface DialogStore {
   selectedTodoId: string | null;
 
   // Если тип диалога update, то необходимо передать selectedTodoId, для получения данных для редактирования
-  openDialog(dialog: 'update', selectedTodoId: string): void;
-
-  // В остальных случаях просто передаем dialog
   openDialog(
     dialog: Exclude<DialogName, 'update'>,
-    selectedTodoId?: string | null,
+    selectedTodoId: string,
   ): void;
+
+  // В остальных случаях просто передаем dialog
+  openDialog(dialog: 'update', selectedTodoId?: string | null): void;
 
   closeDialog: () => void;
 }
